app/requirements.txt
Flask
mysql-connector
gunicorn
celery
redis